requirements.txt
pandas==2.3.3
pip-chill==1.0.3
# psycopg2==2.9.11
pysocks==1.7.1
requests==2.32.5
selenium==4.36.0
fastapi==0.115.5
uvicorn==0.32.0
jinja2
gspread
google-auth
python-dotenv
psycopg2-binary==2.9.11
webdriver-manager
uvicorn